// frontend/frontend/src/app/dashboard/patients/types.tsx
export interface Patient {
    id: number;
    nume: string;
    prenume: string;
    data_nasterii: string;
    gen: string;
    telefon: string;
    email: string;
    adresa: string;
    cnp: string;
    serie_numar_buletin: string;
    cetatenie: string;
    loc_nastere: string;
    contact_urgent_nume: string;
    contact_urgent_telefon: string;
  }
  